Extract the thread slice handling to a re-usable class.
