+++ gapic/src/main/com/google/gapid/perfetto/views/ThreadPanel.java
@@ -29,6 +29,7 @@ import com.google.gapid.perfetto.canvas.RenderContext;
 import com.google.gapid.perfetto.canvas.Size;
 import com.google.gapid.perfetto.models.CpuTrack;
 import com.google.gapid.perfetto.models.Selection.CombiningBuilder;
+import com.google.gapid.perfetto.models.SliceTrack;
 import com.google.gapid.perfetto.models.ThreadTrack;
 
 import org.eclipse.swt.SWT;
@@ -154,12 +155,13 @@ public class ThreadPanel extends TrackPanel implements Selectable {
       }
 
       if (expanded) {
-        for (int i = 0; i < data.starts.length; i++) {
-          long tStart = data.starts[i];
-          long tEnd = data.ends[i];
-          int depth = data.depths[i];
+        SliceTrack.Data slices = data.slices;
+        for (int i = 0; i < slices.starts.length; i++) {
+          long tStart = slices.starts[i];
+          long tEnd = slices.ends[i];
+          int depth = slices.depths[i];
           //String cat = data.categories[i];
-          String title = data.titles[i];
+          String title = slices.titles[i];
           if (tEnd <= visible.start || tStart >= visible.end) {
             continue;
           }
@@ -255,10 +257,11 @@ public class ThreadPanel extends TrackPanel implements Selectable {
       }
     } else if (expanded) {
       depth--;
-      for (int i = 0; i < data.starts.length; i++) {
-        if (data.depths[i] == depth && data.starts[i] <= t && t <= data.ends[i]) {
-          hoveredTitle = data.titles[i];
-          hoveredCategory = data.categories[i];
+      SliceTrack.Data slices = data.slices;
+      for (int i = 0; i < slices.starts.length; i++) {
+        if (slices.depths[i] == depth && slices.starts[i] <= t && t <= slices.ends[i]) {
+          hoveredTitle = slices.titles[i];
+          hoveredCategory = slices.categories[i];
           if (hoveredTitle.isEmpty()) {
             if (hoveredCategory.isEmpty()) {
               return Hover.NONE;
@@ -272,8 +275,8 @@ public class ThreadPanel extends TrackPanel implements Selectable {
               hoveredCategory.isEmpty() ? Size.ZERO : m.measure(hoveredCategory));
           mouseYpos = Math.max(0, Math.min(mouseYpos - (hoveredSize.h - SLICE_HEIGHT) / 2,
               (1 + track.getThread().maxDepth) * SLICE_HEIGHT - hoveredSize.h));
-          long id = data.ids[i];
-          long ts = data.starts[i];
+          long id = slices.ids[i];
+          long ts = slices.starts[i];
 
           return new Hover() {
             @Override
@@ -295,7 +298,8 @@ public class ThreadPanel extends TrackPanel implements Selectable {
             @Override
             public boolean click() {
               if (id != 0) {
-                state.setSelection(ThreadTrack.getSlice(state.getQueryEngine(), id, ts));
+                state.setSelection(SliceTrack.getSlice(
+                    state.getQueryEngine(), SliceTrack.SliceType.Thread, id, ts));
               }
               return false;
             }
@@ -335,9 +339,9 @@ public class ThreadPanel extends TrackPanel implements Selectable {
         endDepth = Integer.MAX_VALUE;
       }
       builder.add(Kind.Thread, transform(
-          ThreadTrack.getSlices(
+          SliceTrack.getThreadSlices(
               state.getQueryEngine(), track.getThread().utid, ts, startDepth, endDepth),
-          ThreadTrack.Slices::new));
+          SliceTrack.Slices::new));
     }
   }
 }
